sua lai hien thi quiz teacher

+++ src/main/java/com/codegym/quizappbackendmodule6/model/dto/QuizResultDTO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,7 +16,7 @@ public class QuizResultDTO {
     private Long id;
     private String userName;
     private LocalDateTime finishTime;
-    private Long score;
+    private BigDecimal score;
     private Long correctAnswers;
     private Long incorrectAnswers;
 }
